Modularize javascript main code
